Start modify ast to use new abstraction
